split v3 python code into separate directory
