+++ python/scripts/copy_pin.py
@@ -8,7 +8,6 @@ sys.path.append(abspath(join(dirname(__file__), '..', 'src')))
 from api_config import ApiConfig
 from access_token import AccessToken
 from oauth_scope import Scope
-from pin import Pin
 
 def main(argv=[]):
     parser = argparse.ArgumentParser(description='Get Pinterest OAuth token')
@@ -20,6 +19,9 @@ def main(argv=[]):
     # get configuration from defaults and/or the environment
     api_config = ApiConfig()
     api_config.verbosity = 2
+
+    # imports that depend on the version of the API
+    from pin import Pin
 
     # Note: It's possible to use the same API configuration with
     # multiple access tokens, so these objects are kept separate.
