+++ python/tests/src/v3/test_user.py
@@ -1,10 +1,10 @@
 import unittest
 import mock
-from src.user import User
+from src.v3.user import User
 
 class UserTest(unittest.TestCase):
-    @mock.patch('src.user.ApiObject.request_data')
-    @mock.patch('src.user.ApiObject.__init__')
+    @mock.patch('src.v3.user.ApiObject.request_data')
+    @mock.patch('src.v3.user.ApiObject.__init__')
     def test_user_get(self, mock_api_object_init, mock_api_object_request_data):
         test_user = User('test_user', 'test_api_uri', 'test_access_token')
         mock_api_object_init.assert_called_once_with('test_api_uri', 'test_access_token')
@@ -14,8 +14,8 @@ class UserTest(unittest.TestCase):
         mock_api_object_request_data.assert_called_once_with('/v3/users/test_user/')
         self.assertEqual(response, 'test_response')
 
-    @mock.patch('src.user.ApiObject.request_data')
-    @mock.patch('src.user.ApiObject.__init__')
+    @mock.patch('src.v3.user.ApiObject.request_data')
+    @mock.patch('src.v3.user.ApiObject.__init__')
     def test_user_get(self, mock_api_object_init, mock_api_object_request_data):
         test_user = User('test_user', 'test_api_uri', 'test_access_token')
         mock_api_object_init.assert_called_once_with('test_api_uri', 'test_access_token')
